Moved category list widget contents to education screen
