--- mobile-app/src/shared/components/category-list-widget/category-list-widget.style.ts
@@ -1,3 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({});
