--- mobile-app/src/shared/components/category-list-widget/index.ts
@@ -1,1 +1,0 @@
-export * from './category-list-widget';
